nus mod unit test, refactor tagtest to subfield
